Restructure view package
