+++ src/main/java/com/thebois/controllers/ColonyController.java
@@ -1,7 +1,7 @@
 package com.thebois.controllers;
 
 import com.thebois.models.world.World;
-import com.thebois.views.ColonyView;
+import com.thebois.views.gameviews.ColonyView;
 
 /**
  * Collects data form the world and sends it to its corresponding view.
