+++ src/main/java/com/thebois/controllers/RoleController.java
@@ -8,8 +8,8 @@ import com.thebois.models.beings.roles.AbstractRole;
 import com.thebois.models.beings.roles.IRoleAllocator;
 import com.thebois.models.beings.roles.RoleFactory;
 import com.thebois.models.beings.roles.RoleType;
-import com.thebois.views.RoleView;
-import com.thebois.views.SpinnerButton;
+import com.thebois.views.infoviews.RoleView;
+import com.thebois.views.infoviews.SpinnerButton;
 
 /**
  * Controls the interactions from the player regarding role allocations.
