+++ src/main/java/com/thebois/views/gameviews/ColonyView.java
@@ -1,4 +1,4 @@
-package com.thebois.views;
+package com.thebois.views.gameviews;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -11,6 +11,7 @@ import com.badlogic.gdx.graphics.g2d.Batch;
 import com.thebois.models.beings.IBeing;
 import com.thebois.models.beings.IBeingGroup;
 import com.thebois.models.beings.roles.RoleType;
+import com.thebois.views.IView;
 
 /**
  * Used to display all pawns.
