+++ src/main/java/com/thebois/views/infoviews/IActorView.java
@@ -1,4 +1,4 @@
-package com.thebois.views;
+package com.thebois.views.infoviews;
 
 import com.badlogic.gdx.scenes.scene2d.Actor;
