+++ src/main/java/com/thebois/views/infoviews/InfoView.java
@@ -1,4 +1,4 @@
-package com.thebois.views;
+package com.thebois.views.infoviews;
 
 import java.util.List;
 
@@ -11,6 +11,8 @@ import com.badlogic.gdx.scenes.scene2d.ui.VerticalGroup;
 import com.badlogic.gdx.scenes.scene2d.ui.WidgetGroup;
 import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
+
+import com.thebois.views.infoviews.IActorView;
 
 /**
  * Displays info about the Colony to the player and allows them to interact with it.
