+++ src/main/java/com/thebois/views/infoviews/InventoryView.java
@@ -1,0 +1,5 @@
+package com.thebois.views.infoviews;
+
+public class InventoryView {
+
+}
