+++ src/main/java/com/thebois/views/infoviews/RoleView.java
@@ -1,4 +1,4 @@
-package com.thebois.views;
+package com.thebois.views.infoviews;
 
 import java.util.AbstractMap;
 import java.util.ArrayList;
