+++ src/main/java/com/thebois/views/infoviews/SpinnerButton.java
@@ -1,4 +1,4 @@
-package com.thebois.views;
+package com.thebois.views.infoviews;
 
 import java.util.ArrayList;
 import java.util.Collection;
